requirements.txt
streamlit
torch
sentence-transformers
faiss-cpu
tokenizers
requests
beautifulsoup4
scikit-learn
transformers
